requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
pyyaml==6.0.1
pandas==2.1.4
numpy==1.24.3
googletrans==4.0.0rc1
